Move files to be under lib/
To avoid namespacing issues move all code inside lib/splitclient-rb/ directory
