+++ lib/splitclient-rb.rb
@@ -1,25 +1,25 @@
 require 'splitclient-rb/version'
 
-require 'exceptions/sdk_blocker_timeout_expired_exception'
-require 'cache/adapters/memory_adapters/map_adapter'
-require 'cache/adapters/memory_adapters/queue_adapter'
-require 'cache/adapters/memory_adapter'
-require 'cache/adapters/redis_adapter'
-require 'cache/repositories/repository'
-require 'cache/repositories/segments_repository'
-require 'cache/repositories/splits_repository'
-require 'cache/repositories/impressions_repository'
-require 'cache/repositories/impressions/memory_repository'
-require 'cache/repositories/impressions/redis_repository'
-require 'cache/repositories/metrics_repository'
-require 'cache/repositories/metrics/memory_repository'
-require 'cache/repositories/metrics/redis_repository'
-require 'cache/senders/impressions_formatter'
-require 'cache/senders/impressions_sender'
-require 'cache/senders/metrics_sender'
-require 'cache/stores/sdk_blocker'
-require 'cache/stores/segment_store'
-require 'cache/stores/split_store'
+require 'splitclient-rb/exceptions/sdk_blocker_timeout_expired_exception'
+require 'splitclient-rb/cache/adapters/memory_adapters/map_adapter'
+require 'splitclient-rb/cache/adapters/memory_adapters/queue_adapter'
+require 'splitclient-rb/cache/adapters/memory_adapter'
+require 'splitclient-rb/cache/adapters/redis_adapter'
+require 'splitclient-rb/cache/repositories/repository'
+require 'splitclient-rb/cache/repositories/segments_repository'
+require 'splitclient-rb/cache/repositories/splits_repository'
+require 'splitclient-rb/cache/repositories/impressions_repository'
+require 'splitclient-rb/cache/repositories/impressions/memory_repository'
+require 'splitclient-rb/cache/repositories/impressions/redis_repository'
+require 'splitclient-rb/cache/repositories/metrics_repository'
+require 'splitclient-rb/cache/repositories/metrics/memory_repository'
+require 'splitclient-rb/cache/repositories/metrics/redis_repository'
+require 'splitclient-rb/cache/senders/impressions_formatter'
+require 'splitclient-rb/cache/senders/impressions_sender'
+require 'splitclient-rb/cache/senders/metrics_sender'
+require 'splitclient-rb/cache/stores/sdk_blocker'
+require 'splitclient-rb/cache/stores/segment_store'
+require 'splitclient-rb/cache/stores/split_store'
 
 require 'splitclient-rb/localhost_utils'
 require 'splitclient-rb/clients/localhost_split_client'
@@ -31,41 +31,41 @@ require 'splitclient-rb/split_factory_builder'
 require 'splitclient-rb/localhost_split_factory'
 require 'splitclient-rb/split_config'
 
-require 'engine/api/client'
-require 'engine/api/impressions'
-require 'engine/api/metrics'
-require 'engine/api/segments'
-require 'engine/api/splits'
-require 'engine/parser/condition'
-require 'engine/parser/partition'
-require 'engine/parser/split_adapter'
-require 'engine/parser/evaluator'
-require 'engine/matchers/combiners'
-require 'engine/matchers/combining_matcher'
-require 'engine/matchers/all_keys_matcher'
-require 'engine/matchers/negation_matcher'
-require 'engine/matchers/user_defined_segment_matcher'
-require 'engine/matchers/whitelist_matcher'
-require 'engine/matchers/equal_to_matcher'
-require 'engine/matchers/greater_than_or_equal_to_matcher'
-require 'engine/matchers/less_than_or_equal_to_matcher'
-require 'engine/matchers/between_matcher'
-require 'engine/matchers/set_matcher'
-require 'engine/matchers/part_of_set_matcher'
-require 'engine/matchers/equal_to_set_matcher'
-require 'engine/matchers/contains_any_matcher'
-require 'engine/matchers/contains_all_matcher'
-require 'engine/matchers/starts_with_matcher'
-require 'engine/matchers/ends_with_matcher'
-require 'engine/matchers/contains_matcher'
-require 'engine/matchers/dependency_matcher'
-require 'engine/evaluator/splitter'
-require 'engine/metrics/metrics'
-require 'engine/metrics/binary_search_latency_tracker'
-require 'engine/models/split'
-require 'engine/models/label'
-require 'engine/models/treatment'
-require 'splitclient-rb_utilitites'
+require 'splitclient-rb/engine/api/client'
+require 'splitclient-rb/engine/api/impressions'
+require 'splitclient-rb/engine/api/metrics'
+require 'splitclient-rb/engine/api/segments'
+require 'splitclient-rb/engine/api/splits'
+require 'splitclient-rb/engine/parser/condition'
+require 'splitclient-rb/engine/parser/partition'
+require 'splitclient-rb/engine/parser/split_adapter'
+require 'splitclient-rb/engine/parser/evaluator'
+require 'splitclient-rb/engine/matchers/combiners'
+require 'splitclient-rb/engine/matchers/combining_matcher'
+require 'splitclient-rb/engine/matchers/all_keys_matcher'
+require 'splitclient-rb/engine/matchers/negation_matcher'
+require 'splitclient-rb/engine/matchers/user_defined_segment_matcher'
+require 'splitclient-rb/engine/matchers/whitelist_matcher'
+require 'splitclient-rb/engine/matchers/equal_to_matcher'
+require 'splitclient-rb/engine/matchers/greater_than_or_equal_to_matcher'
+require 'splitclient-rb/engine/matchers/less_than_or_equal_to_matcher'
+require 'splitclient-rb/engine/matchers/between_matcher'
+require 'splitclient-rb/engine/matchers/set_matcher'
+require 'splitclient-rb/engine/matchers/part_of_set_matcher'
+require 'splitclient-rb/engine/matchers/equal_to_set_matcher'
+require 'splitclient-rb/engine/matchers/contains_any_matcher'
+require 'splitclient-rb/engine/matchers/contains_all_matcher'
+require 'splitclient-rb/engine/matchers/starts_with_matcher'
+require 'splitclient-rb/engine/matchers/ends_with_matcher'
+require 'splitclient-rb/engine/matchers/contains_matcher'
+require 'splitclient-rb/engine/matchers/dependency_matcher'
+require 'splitclient-rb/engine/evaluator/splitter'
+require 'splitclient-rb/engine/metrics/metrics'
+require 'splitclient-rb/engine/metrics/binary_search_latency_tracker'
+require 'splitclient-rb/engine/models/split'
+require 'splitclient-rb/engine/models/label'
+require 'splitclient-rb/engine/models/treatment'
+require 'splitclient-rb/utilitites'
 
 module SplitIoClient
   def self.root
