+++ lib/splitclient-rb/cache/adapters/memory_adapter.rb
