+++ lib/splitclient-rb/cache/adapters/memory_adapters/map_adapter.rb
