+++ lib/splitclient-rb/cache/adapters/redis_adapter.rb
