+++ lib/splitclient-rb/cache/senders/impressions_formatter.rb
