+++ lib/splitclient-rb/cache/stores/segment_store.rb
