+++ lib/splitclient-rb/engine/api/metrics.rb
