+++ lib/splitclient-rb/engine/api/segments.rb
