+++ lib/splitclient-rb/engine/evaluator/splitter.rb
