+++ lib/splitclient-rb/engine/matchers/ends_with_matcher.rb
