+++ lib/splitclient-rb/engine/matchers/equal_to_set_matcher.rb
