+++ lib/splitclient-rb/engine/matchers/greater_than_or_equal_to_matcher.rb
@@ -38,7 +38,7 @@ module SplitIoClient
           return value
         when "DATETIME"
           value = value/1000 if is_sdk_data # sdk returns already miliseconds, turning to seconds to do a correct zero_our
-          return ::Utilities.to_milis_zero_out_from_seconds value
+          return SplitIoClient::Utilities.to_milis_zero_out_from_seconds value
         else
           @logger.error('Invalid data type')
       end
