+++ lib/splitclient-rb/engine/matchers/negation_matcher.rb
