+++ lib/splitclient-rb/engine/matchers/set_matcher.rb
