+++ lib/splitclient-rb/engine/matchers/starts_with_matcher.rb
