+++ lib/splitclient-rb/engine/matchers/whitelist_matcher.rb
