+++ lib/splitclient-rb/engine/metrics/binary_search_latency_tracker.rb
