+++ lib/splitclient-rb/engine/models/split.rb
