+++ lib/splitclient-rb/engine/parser/condition.rb
