+++ lib/splitclient-rb/engine/parser/evaluator.rb
