+++ lib/splitclient-rb/engine/parser/split_adapter.rb
